Collapsable sections & toolbar plugins
* Collapsable sections plugin
Allows sections of the page to be hidden until a button is pressed.
This is used for hiding the answers to exercises in
the workshop.

* Toolbar plugin
Adds icons to the existing toolbar at the top right of each page.  Added
icon links for
- Gitter.im chat room for Practicalli Github Organisation
- Github Practicalli organisation home
- StackOverflow search using title of current page

Also added file patterns to ignore in the .bookignore